// backend/src/components/product/interfaces/OrderProduct.ts
// Interface automatically generated by schemas-to-ts

import { ProductColor } from '../../../api/product-color/content-types/product-color/product-color';
import { ProductSize } from '../../../api/product-size/content-types/product-size/product-size';
import { Product } from '../../../api/product/content-types/product/product';
import { ProductColor_Plain } from '../../../api/product-color/content-types/product-color/product-color';
import { ProductSize_Plain } from '../../../api/product-size/content-types/product-size/product-size';
import { Product_Plain } from '../../../api/product/content-types/product/product';

export interface OrderProduct {
  product_color?: { data: ProductColor };
  product_size?: { data: ProductSize };
  quantity: number;
  product?: { data: Product };
}
export interface OrderProduct_Plain {
  product_color?: ProductColor_Plain;
  product_size?: ProductSize_Plain;
  quantity: number;
  product?: Product_Plain;
}

export interface OrderProduct_NoRelations {
  product_color?: number;
  product_size?: number;
  quantity: number;
  product?: number;
}